% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USANatality2019.R
\docType{data}
\name{USANatality2019}
\alias{USANatality2019}
\title{Birth (natality) data from the United States of America in 2019}
\format{
The data consists of 3,757,582 rows with 238 observations per row
The columns are described in \url{ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/Dataset_Documentation/DVS/natality/UserGuide2019-508.pdf}
}
\source{
\url{ftp://ftp.cdc.gov/pub/Health_Statistics/NCHS/Datasets/DVS/natality/Nat2019us.zip}
}
\usage{
USANatality2019
}
\description{
Birth (natality) data from the United States of America in 2019
}
\keyword{datasets}
